refactor: driver 패키지 하위의 엔티티들을 entity 패키지로 이동
