+++ BE/src/main/java/ifive/idrop/domain/driver/dto/DriverResponse.java
@@ -1,6 +1,6 @@
 package ifive.idrop.domain.driver.dto;
 
-import ifive.idrop.domain.driver.Driver;
+import ifive.idrop.domain.driver.entity.Driver;
 import lombok.Builder;
 import lombok.Getter;
