+++ BE/src/main/java/ifive/idrop/domain/driver/entity/WorkSchedule.java
@@ -1,4 +1,4 @@
-package ifive.idrop.domain.driver;
+package ifive.idrop.domain.driver.entity;
 
 import ifive.idrop.common.enums.Day;
 import jakarta.persistence.*;
