+++ BE/src/main/java/ifive/idrop/domain/pickup/controller/PickUpController.java
@@ -2,7 +2,7 @@ package ifive.idrop.domain.pickup.controller;
 
 import ifive.idrop.domain.auth.resolver.Login;
 import ifive.idrop.common.dto.BaseResponse;
-import ifive.idrop.domain.driver.Driver;
+import ifive.idrop.domain.driver.entity.Driver;
 import ifive.idrop.domain.pickup.PickUpHistory;
 import ifive.idrop.common.exception.BusinessException;
 import ifive.idrop.common.exception.ErrorCode;
